Extracted the syncing of scrolls into a hook
