+++ src/lib/components/tablev2/MultiSelectableContent.tsx
@@ -21,6 +21,7 @@ import {
   TableVariantType,
 } from './TableUtils';
 import { useTableScrollbar, VirtualizedRows } from './TableCommon';
+import useSyncedScroll from './useSyncedScroll';
 
 const translations = {
   en: {
@@ -107,33 +108,7 @@ export const MultiSelectableContent = <
     return index;
   };
 
-  const headerRef = useRef<HTMLDivElement | null>(null);
-  const bodyRef = useRef<FixedSizeList<Row<DATA_ROW>[]> | null>(null);
-  const currentHeadRef = !!headerRef.current;
-  const currentBodyRef = !!bodyRef.current;
-  useEffect(() => {
-    if (bodyRef.current && headerRef.current) {
-      const listener = (event: Event) => {
-        headerRef.current.scrollTo({
-          left: (event.target as HTMLDivElement).scrollLeft,
-          top: 0,
-        });
-      };
-      /*
-      We intentionally use _outerRef prop here despite the fact that it is 
-      internal use only and not typed, as it is the only way for us to access to the scrollable element
-      */
-      //@ts-expect-error
-      (bodyRef.current._outerRef as HTMLDivElement).addEventListener(
-        'scroll',
-        listener,
-      );
-      return () => {
-        //@ts-expect-error
-        bodyRef.current._outerRef.removeEventListener('scroll', listener);
-      };
-    }
-  }, [currentHeadRef, currentBodyRef]);
+  const { bodyRef, headerRef } = useSyncedScroll<DATA_ROW>();
 
   const RenderRow = memo(({ index, style }: RenderRowType) => {
     const row = rows[index];
